+++ bootstrap/http/start.php
@@ -52,4 +52,4 @@ $application->start(function () use ($application, $container) {
  * Shut her down
  * ----------------------------------------------------------
  */
-$application->shutdown();
+$application->shutDown();