+++ configs/application.php
@@ -1,7 +1,6 @@
 <?php
 use Opulence\Applications\Application;
 use Opulence\Applications\Environments\Environment;
-use Opulence\Applications\Paths;
 use Opulence\Applications\Tasks\Dispatchers\IDispatcher as ITaskDispatcher;
 use Opulence\Bootstrappers\ApplicationBinder;
 use Opulence\Bootstrappers\BootstrapperRegistry;
@@ -10,6 +9,7 @@ use Opulence\Bootstrappers\Caching\ICache;
 use Opulence\Bootstrappers\Dispatchers\Dispatcher as BootstrapperDispatcher;
 use Opulence\Bootstrappers\Dispatchers\IDispatcher as IBootstrapperDispatcher;
 use Opulence\Bootstrappers\IBootstrapperRegistry;
+use Opulence\Bootstrappers\Paths;
 use Opulence\Ioc\IContainer;
 
 /**